+++ packages/ulangi-common/src/converters/index.ts
@@ -6,7 +6,6 @@
  */
 
 export { LanguageConverter } from './LanguageConverter';
-export { WordClassConverter } from './WordClassConverter';
 export { PublicVocabularyConverter } from './PublicVocabularyConverter';
 export { TranslationConverter } from './TranslationConverter';
 export { DictionaryEntryConverter } from './DictionaryEntryConverter';