+++ scripts/run_postbootstrap_all_packages.sh
@@ -8,6 +8,6 @@ cd packages
 echo "Executing postbootstrap script of each package..."
 for d in */ ; do
   cd ${PROJECT_DIR}/packages/$d
-  npm run postbootstrap --if-present
+  npm run postbootstrap --if-present || exit 1
 done
 echo "Executed postbootstrap script of each package."