+++ packages/ulangi-local-database/CHANGELOG.md
@@ -3,6 +3,45 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.11.5](https://github.com/ulangi/ulangi/compare/v3.11.4...v3.11.5) (2020-04-12)
+
+**Note:** Version bump only for package @ulangi/ulangi-local-database
+
+
+
+
+
+## [3.11.3](https://github.com/ulangi/ulangi/compare/v3.11.2...v3.11.3) (2020-03-11)
+
+
+### Bug Fixes
+
+* use custom build sqlite for Android ([26d0fa7](https://github.com/ulangi/ulangi/commit/26d0fa7))
+* use journal_mode DELETE by default ([2af60ab](https://github.com/ulangi/ulangi/commit/2af60ab))
+
+
+
+
+
+## [3.11.2](https://github.com/ulangi/ulangi/compare/v3.11.1...v3.11.2) (2020-03-10)
+
+
+### Bug Fixes
+
+* download incompatible sets/vocabulary on remote changes ([b17f004](https://github.com/ulangi/ulangi/commit/b17f004))
+
+
+
+
+
+## [3.11.1](https://github.com/ulangi/ulangi/compare/v3.11.0...v3.11.1) (2020-03-08)
+
+**Note:** Version bump only for package @ulangi/ulangi-local-database
+
+
+
+
+
 # [3.11.0](https://github.com/minhloi/ulangi/compare/v3.10.1...v3.11.0) (2020-03-07)
 
 **Note:** Version bump only for package @ulangi/ulangi-local-database