+++ packages/ulangi-mobile/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.11.4](https://github.com/ulangi/ulangi/compare/v3.11.3...v3.11.4) (2020-03-13)
+
+
+### Bug Fixes
+
+* a minor bug fix ([7b32a80](https://github.com/ulangi/ulangi/commit/7b32a80))
+
+
+
+
+
+## [3.11.3](https://github.com/ulangi/ulangi/compare/v3.11.2...v3.11.3) (2020-03-11)
+
+
+### Bug Fixes
+
+* minor changes for Google Sheets add-on ([cc03342](https://github.com/ulangi/ulangi/commit/cc03342))
+* use custom build sqlite for Android ([26d0fa7](https://github.com/ulangi/ulangi/commit/26d0fa7))
+
+
+
+
+
+## [3.11.2](https://github.com/ulangi/ulangi/compare/v3.11.1...v3.11.2) (2020-03-10)
+
+
+### Bug Fixes
+
+* don't need to record error before throw a saga error ([3d17cdf](https://github.com/ulangi/ulangi/commit/3d17cdf))
+* make sure MainActivity only created once ([4c577bc](https://github.com/ulangi/ulangi/commit/4c577bc))
+
+
+
+
+
+## [3.11.1](https://github.com/ulangi/ulangi/compare/v3.11.0...v3.11.1) (2020-03-08)
+
+**Note:** Version bump only for package @ulangi/ulangi-mobile
+
+
+
+
+
 # [3.11.0](https://github.com/minhloi/ulangi/compare/v3.10.1...v3.11.0) (2020-03-07)